requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
openai==1.109.1
requests==2.32.5
python-dotenv==1.2.1
langchain==0.2.17
langchain-openai==0.1.25
beautifulsoup4==4.14.3
pypdf==6.6.2
langsmith==0.1.147
pymilvus==2.6.8
sentence-transformers==5.2.2
guardrails-ai==0.5.6
pydantic==2.12.5
pydantic-settings==2.12.0
click==8.1.7
typer==0.12.5